+++ src/main/java/com/projectmatching/app/service/user/UserService.java
@@ -22,7 +22,7 @@ public interface UserService {
     Long updateUser(UserDto userDto);
 
     //특정 유저 좋아요 누르기
-    Long addLiking(UserDetailsImpl userDetails, int userId);
+    Long addLiking(UserDetailsImpl userDetails, long userId);
